19: Moving ProjectsListItem component to folder
